+++ sendhooks/sender/sender_test.go
@@ -2,7 +2,7 @@ package sender
 
 /*
 This package contains mostly test functions for the utils used to send webhooks. If these utils works,
-we can ensure that the send webhook function will function normally too. For a whole test suite for the sender function,
+we can ensure that the send sendhooks function will function normally too. For a whole test suite for the sender function,
 check out the webhook_test.go file.
 */
 
@@ -10,8 +10,8 @@ import (
 	"bytes"
 	"io"
 	"net/http"
+	"sendhooks/redis"
 	"testing"
-	"webhook/redis"
 
 	"github.com/stretchr/testify/assert"
 )